# Conflicts:
#	core/src/com/mygdx/chalmersdefense/model/Model.java
#	core/src/com/mygdx/chalmersdefense/model/towers/ChemistTower.java
#	core/src/com/mygdx/chalmersdefense/model/towers/EcoTower.java
#	core/src/com/mygdx/chalmersdefense/model/towers/ElectroTower.java
#	core/src/com/mygdx/chalmersdefense/model/towers/HackerTower.java
#	core/src/com/mygdx/chalmersdefense/model/towers/MeckTower.java
#	core/src/com/mygdx/chalmersdefense/model/towers/SmurfTower.java
#	core/src/com/mygdx/chalmersdefense/model/towers/Tower.java

+++ core/src/com/mygdx/chalmersdefense/views/GameScreen.java
@@ -189,7 +189,7 @@ public class GameScreen extends AbstractScreen implements Screen {
             Sprite towerSprite = spriteMap.get(tower.getSpriteKey());
 
             towerSprite.setPosition(tower.getPosX(), tower.getPosY());
-            towerSprite.setRotation(tower.getAngle());
+            towerSprite.setRotation((float) tower.getAngle());
 
 
             if(!tower.isPlaced() && !tower.getCollision()){
